add wishlist & manage events delete

+++ src/models/events.models.js
@@ -104,9 +104,18 @@ exports.updateManageEvents = async function(id, data){
 
 exports.destroy = async function(id){
     const query = `
-      DELETE FROM "${table}" WHERE "id"=$1 RETURNING *
-    `
+    DELETE FROM "${table}" WHERE "id"=$1 RETURNING *
+  `
     const values = [id]
+    const {rows} = await db.query(query, values)
+    return rows[0]
+}
+
+exports.destroyByIdAndUserId = async function(id, createdBy){
+    const query = `
+  DELETE FROM "${table}" WHERE "id"=$1 AND "createdBy" = $2 RETURNING *
+`
+    const values = [id, createdBy]
     const {rows} = await db.query(query, values)
     return rows[0]
 }
